Refactored user view tests into admin and ajax files. Added comments for process of deleting a user.
